+++ CHANGELOG.md
@@ -6,6 +6,9 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## Unreleased
+### Added 
+- Add filter by ids to public accounts [#744](https://github.com/rokwire/core-building-block/issues/744)
+
 ## [1.45.2] - 2024-18-12
 ### Fixed
 - Fix Case-Insensitive Sorting [#741](https://github.com/rokwire/core-building-block/issues/741)